admin page without edit invoice, edit my profile temp client

+++ app/Views/admin/invoice/index.php
@@ -38,14 +38,13 @@
                                     <td>
                                         <div class="obj-action">
                                             <div class="ac">
-
                                                 <a href="<?= base_url() ?>/admin/Invoices/detail?id=<?= $row['id'] ?>" data-toggle="tooltip" data-placement="bottom" title="Detail"><i class="fas fa-info-circle"></i></a>
                                             </div>
                                             <div class="ac">
                                                 <a href="<?= base_url() ?>/admin/Invoices/edit?id=<?= $row['id'] ?>" data-toggle="tooltip" data-placement="bottom" title="Edit"><i class="far fa-edit"></i> </a>
                                             </div>
                                             <div class="ac">
-                                                <a href="<?= base_url() ?>/admin/Invoices/delete?id=<?= $row['id'] ?>" data-toggle="tooltip" data-placement="bottom" title="Delete"><i class="far fa-trash-alt"></i></a>
+                                                <a href="<?= base_url() ?>/admin/Invoices/delete?id=<?= $row['id'] ?>" onclick="return confirm('Are you sure?');" data-toggle="tooltip" data-placement="bottom" title="Delete"><i class="far fa-trash-alt"></i></a>
                                             </div>
 
                                         </div>
